Rename controller account -> accounts. Adjust imports. Small test style refactor
