+++ framework-plugins/lisk-framework-http-api-plugin/src/controllers/accounts.ts
